+++ tests/testthat/test_heterogenous_calculations.R
@@ -62,8 +62,8 @@ test_that("Heterogenous model object is constructed correctly",{
 })
 
 test_that("Kernel matrices are extracted correctly", {
-  expect_equal(K, get_kernel(mod, 'row'))
-  expect_equal(G, get_kernel(mod, 'column'))
+  expect_equal(K, get_kernelmatrix(mod, 'row'))
+  expect_equal(G, get_kernelmatrix(mod, 'column'))
 })
 
 # Check label matching